rename metadata interfaces to metadata structs
