+++ scos_actions/actions/acquire_single_freq_tdomain_iq.py
@@ -38,8 +38,8 @@ from numpy import complex64
 from scos_actions import utils
 from scos_actions.actions.interfaces.measurement_action import MeasurementAction
 from scos_actions.hardware.mocks.mock_gps import MockGPS
-from scos_actions.metadata.interfaces import ntia_sensor
-from scos_actions.metadata.interfaces.capture import CaptureSegment
+from scos_actions.metadata.structs import ntia_sensor
+from scos_actions.metadata.structs.capture import CaptureSegment
 from scos_actions.utils import get_parameter
 
 logger = logging.getLogger(__name__)
