+++ scos_actions/metadata/structs/ntia_algorithm.py
