+++ scos_actions/metadata/structs/ntia_diagnostics.py
