+++ scos_actions/metadata/structs/ntia_environment.py
